Move numberToFraction into Fraction class
